rename move and fix composables document composables
